Add new algorithm.hpp file.
* Move emulation of std::lexicographical_compare_three_way to its own
  algorithm.hpp file.
